+++ python/math/setup.cfg
@@ -1,0 +1,25 @@
+[metadata]
+name = inertialsense_math
+version = 0.0.3
+author = Inertial Sense Build Bot
+author_email = steven+pypi.org@inertialsense.com
+description = Inertial Sense Math Functions
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/inertialsense/inertial-sense-sdk
+project_urls = 
+	Bug Tracker = https://github.com/inertialsense/inertial-sense-sdk/issues
+classifiers = 
+	Programming Language :: Python :: 3
+	License :: OSI Approved :: MIT License
+	Operating System :: OS Independent
+
+[options]
+package_dir = 
+	= src
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = src
+
